Move the pooling parameter checks to the op layer
``ops/pool.py`` took ``_normalize_pool_dims`` and ``validate_pool_params`` from
``kernels/pool/common.py``. Validation and normalization run for every target, so
they belong on the side that does them, not in the package that holds one target's
kernels.

``pool_output_dim`` stays with the kernels: all six of them compute their own output
extents with it, and this layer needs the same arithmetic for
``_infer_output_shapes``. A neutral home would mean a new kernels-to-manifest
dependency, which is not this change's call to make.
